requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
django-jazzmin==2.6.1
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pydotplus==2.0.2
pygraphviz==1.12
pyparsing==3.1.2
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
